Client envois pos en continu

+++ test/core/src/Util/Coord.java
@@ -2,15 +2,15 @@ package Util;
 
 public class Coord {
 
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public Coord() {
         x = 0;
         y = 0;
     }
 
-    public Coord(int x, int y) {
+    public Coord(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -32,19 +32,19 @@ public class Coord {
         --x;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
